+++ api/database/migrations/2020_09_24_112843_create_majestic_table.php
@@ -16,14 +16,14 @@ class CreateMajesticTable extends Migration
         Schema::create('majestic', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('platform_id');
-            $table->integer('external_backlinks')->nullable();
-            $table->integer('external_gov')->nullable();
-            $table->integer('external_edu')->nullable();
-            $table->integer('tf')->nullable();
-            $table->integer('cf')->nullable();
-            $table->integer('refd')->nullable();
-            $table->integer('refd_edu')->nullable();
-            $table->integer('refd_gov')->nullable();
+            $table->unsignedBigInteger('external_backlinks')->nullable();
+            $table->unsignedBigInteger('external_gov')->nullable();
+            $table->unsignedBigInteger('external_edu')->nullable();
+            $table->unsignedBigInteger('tf')->nullable();
+            $table->unsignedBigInteger('cf')->nullable();
+            $table->unsignedBigInteger('refd')->nullable();
+            $table->unsignedBigInteger('refd_edu')->nullable();
+            $table->unsignedBigInteger('refd_gov')->nullable();
 
             $table
                 ->foreign('platform_id')
